fix - Exception generate

+++ src/main/java/com/ainapapy/aigle/services/UserService.java
@@ -1,5 +1,6 @@
 package com.ainapapy.aigle.services;
 
+import com.ainapapy.aigle.exceptions.UserNotFoundException;
 import com.ainapapy.aigle.models.User;
 import com.ainapapy.aigle.models.convertors.UserConvertor;
 import com.ainapapy.aigle.models.dto.UserDTO;
@@ -61,7 +62,7 @@ public class UserService {
             user.setRole(userConvertor.convertToRoleEnum(dto.getRole()));
             return userConvertor.convertToDTO(this.saveUser(user));
         }
-        return null;
+        throw new UserNotFoundException("User for Put ID: " + id + " not found");
     }
     
     public UserDTO PatchUser(Long id, UserDTO dto) {
@@ -80,11 +81,15 @@ public class UserService {
             
             return userConvertor.convertToDTO(this.saveUser(user));
         }
-        return null;
+        throw new UserNotFoundException("User for Patch ID: " + id + " not found");
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> userToDelete = this.getUserById(id);
+        if (userToDelete.isPresent()) {
+            userRepository.deleteById(id);
+        }
+        throw new UserNotFoundException("User for Delete ID: " + id + " not found");
     }
